:arrow_right: DataMask template class to its own file
